Refactor files and rollback deletions to fulfill updated user story 26

+++ app/controllers/orders_controller.rb
@@ -1,4 +1,6 @@
 class OrdersController <ApplicationController
+  def new
+  end
 
   def index
     @user = current_user
@@ -10,27 +12,28 @@ class OrdersController <ApplicationController
 
   def create
     user = User.find(params[:id])
-    order = user.orders.create(order_params)
-    cart.items.each do |item,quantity|
-      order.item_orders.create({
-        item: item,
-        quantity: quantity,
-        price: item.price
-        })
+    order = user.orders.new(order_params)
+    if order.save
+      cart.items.each do |item,quantity|
+        order.item_orders.create({
+          item: item,
+          quantity: quantity,
+          price: item.price
+          })
+      end
+      session.delete(:cart)
+      flash[:order_created] = 'Order ho! Ye successfully placed yer order. Make way in yer hold for loot!'
+      redirect_to "/profile/orders"
+    else
+      flash[:notice] = "Please complete address form to create an order."
+      render :new
     end
-    session.delete(:cart)
-    flash[:order_created] = 'Order ho! Ye successfully placed yer order. Make way in yer hold for loot!'
-    redirect_to "/profile/orders"
   end
 
 
   private
 
   def order_params
-    {name: current_user.name,
-    address: current_user.address,
-    state: current_user.state,
-    city: current_user.city,
-    zip: current_user.zip}
+    params.permit(:name, :address, :city, :state, :zip)
   end
 end
